pyproject.toml
[project]
name = "scada-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Fabian",email = "heromfabian@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "fastapi-users[sqlalchemy] (>=15.0.3,<16.0.0)",
    "sqlmodel (>=0.0.31,<0.0.32)",
    "aiomqtt (>=2.5.0,<3.0.0)",
    "pymodbus (>=3.11.4,<4.0.0)",
    "asyncua (>=1.1.8,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
